+++ zebrad/src/components.rs
@@ -10,6 +10,7 @@ pub mod inbound;
 #[allow(missing_docs)]
 pub mod mempool;
 pub mod metrics;
+pub mod notify;
 #[allow(missing_docs)]
 pub mod sync;
 #[allow(missing_docs)]